+++ tests/Phan/PhanTest8.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phan\Tests;
+
+/**
+ * Runs tests/files/src/0800-0899
+ */
+class PhanTest8 extends PhanTestRange
+{
+    public const START_RANGE = '0800';
+    public const END_RANGE   = '0900';
+}